requirements.txt
alembic==1.6.5
anyio==3.3.0
apache-airflow==2.1.2
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-sqlite==2.0.0
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.9.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
blinker==1.4
cattrs==1.5.0
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.7
defusedxml==0.7.1
dill==0.3.4
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
graphviz==0.17
gunicorn==20.1.0
h11==0.12.0
html5lib==1.1
httpcore==0.13.6
httpx==0.19.0
idna==2.10
importlib-resources==1.5.0
inflection==0.5.1
install==1.3.4
iso8601==0.1.16
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
kafka-python==2.0.2
lazy-object-proxy==1.6.0
lockfile==0.12.2
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
numpy==1.21.2
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
pandas==1.3.2
paramiko==2.7.2
pendulum==2.1.2
prison==0.1.3
psutil==5.8.0
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pymongo==3.11.4
PyNaCl==1.4.0
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
requests==2.25.1
rfc3986==1.5.0
rich==10.7.0
selenium==3.141.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.2.1
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
sshtunnel==0.4.0
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==6.2.0
termcolor==1.1.0
text-unidecode==1.3
unicodecsv==0.14.1
urllib3==1.26.4
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
